requirements.txt
rich>=12.0.0
click>=8.0.0
requests>=2.28.0
jinja2>=3.1.0
dnspython>=2.3.0
aiohttp>=3.8.0
python-nmap>=0.7.1
google-generativeai>=0.3.0
